Move osc control element creation to Views file
